Warn instead of failing in double hashed client
Rationale is that double hashed client has a larger surface for failure comparing to the regular client. It does multiple network roundtrips as well as a number of decryption operations in order to fullfill a single lookup. We should be striving for returning a consistent result to the user even if some parts of the query might fail. Logging instead of failing will also allow us to do finer grained debugging in dhfind if necessary. The next step will be to enable streaming to reduce time to the first result.