Return content not found error when HTTP publisher responds with 404
The entry chunks for which content is not found are skipped by the
indexer. In HTTP publisher, non-existing content returns 404. However,
the error returned by the syncer does not allow the "content not found"
error to bubble up to ingest pipeline. Therefore, the indexer cannot
skip such CIDs from HTTP publishers.

In HTTP Syncer check the response status code and if 404 return relevant
errors for upstream error handling.

Relates to:
 - https://github.com/ipni/storetheindex/issues/1771