DHClient can have a dhstore URL or API  (#71)
* DHClient can have a dhstore URL or API

DHashClient can be given a dhstore API interface when the client within dhstore, This is so that dhstore can give it a set of internal functions to call to do dh multihash and metadata lookups. Otherwise, a client that is not in dhstore uses the dhstore URL to perfrom remote dhstore lookups.