Fix logging "encrypted metadata too short" when metadata no found
When a multihash no longer maps to a metadata, because that metadata was deleted, a NotFound error is returned in the HTTP response in the normal dhstore metadata lookup. With dhfind, the FindMetadata is called directly so that NotFound error is not returned and instead nil data is received. When decryption is done on the nil data, this logs the error. No harm is done, since the logic takes the same path as handling a NotFound error, but the error should not be logged as this is a normal occurrance.

This PR fixes this by handling a not found the same whether it comes from an HTTP response or from value returned by direct call.