Decoupe announce Senders from Publishers
There is no strong reason to associate a set of announce Senders with a Publisher. The only association is that a Sender is often used to after updating the Sublisher's head CID. By separating these, it allows more flexibility in where in handling serving an advertisement chain is done and where sending advertisement announcements is done.